% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_zint.R
\name{check_zint}
\alias{check_zint}
\title{Check Zint}
\usage{
check_zint()
}
\value{
a numeric value of 0 or 1. O if \code{zint} is not available and 1 if it is
}
\description{
Check that zint is installed
}
\keyword{internal}
