% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_address_barcode.R
\name{create_address_barcode}
\alias{create_address_barcode}
\title{Create Address Barcode}
\usage{
create_address_barcode(address = "CarwynJamesBuilding")
}
\description{
Create Address Barcode
}
