% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directory.R
\name{create_directory}
\alias{create_directory}
\title{Create Temporary Directory}
\usage{
create_directory(filePath)
}
\arguments{
\item{filePath}{A valid directory path where barcodes are saved}
}
\description{
Create Temporary Directory
}
