% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_id_barcode.R
\name{create_id_barcode}
\alias{create_id_barcode}
\title{Create ID Barcode}
\usage{
create_id_barcode(id)
}
\arguments{
\item{id}{A character string of participant ID}
}
\description{
This functions invokes a system command to \code{Zint} and outputs a \strong{Code 128} barcode.
}
\details{
Barcodes are saved to the directory which has been created using \code{create_directory}.
}
