% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_barcodes.R
\name{generate_barcodes}
\alias{generate_barcodes}
\title{Generate Barcodes}
\usage{
generate_barcodes(
  filePath,
  ID,
  DOB,
  GENDER,
  ADDRESS,
  TYPE = c("FORM", "LOG", "FBC"),
  pathRename = FALSE
)
}
\arguments{
\item{filePath}{A valid directory path where barcodes are saved}

\item{ID}{A character string of participant ID}

\item{DOB}{A character string of Date-of Birth in the following format; \strong{DD/MM/YYYY}}

\item{GENDER}{A character of either M or F (for Male and Female respectively)}

\item{ADDRESS}{A character string with no white space which is used as the sample address}

\item{TYPE}{A character vector of any of the following options
\itemize{
\item \strong{FORM} : Transfer Form
\item \strong{LOG} : Log Book
\item \strong{FBC} : Full Blood Count
\item \strong{CHEM21} : Chemistry-21 Panel (LFT, U&E)
\item \strong{VITD} : Vitamin-D
\item \strong{GLUCOSE} : Glucose
\item \strong{HBA1C} : Glycated Haemoglobin (HbA1c)
\item \strong{INSULIN} : Insulin
}}

\item{pathRename}{logical; if \code{TRUE} then the directory where barcodes are saved to will be named \code{ID} + \code{Sys.Date()}. IF \code{FALSE} (default)
then the original UID code will be used for directory naming.}
}
\description{
This is the main worker function of the whole package. \code{generate_barcodes} wraps up the barcode creation and arrangement and
saves them in the specified directory.
}
