% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_width_resize.R
\name{image_width_resize}
\alias{image_width_resize}
\title{Resize Image Width}
\usage{
image_width_resize(barcode_image)
}
\arguments{
\item{barcode_imaage}{a \code{magick-image} of barcode that has read in using \code{magick::image_read()}}
}
\value{
a character string of formated width and heght (ie, '93\%x100\%)
}
\description{
Generate the values needed for \code{magick::image_resize()} to scale barcodes width relatve to the DOB barcode
}
\keyword{internal}
