% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_location.R
\name{text_location}
\alias{text_location}
\title{Centre Barcode Text}
\usage{
text_location(x)
}
\arguments{
\item{x}{a character string of the bar code label}
}
\value{
a character string of the formatted location values (ie, '+80+35)
}
\description{
Create the location coordinates to ensure bar code text is centred when using \code{magick::image_annotate} in \code{stack_barcodes}
}
\keyword{internal}
