% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_address_barcode.R
\name{create_address_barcode}
\alias{create_address_barcode}
\title{Create Address Barcode}
\usage{
create_address_barcode(address = "CarwynJamesBuilding")
}
\arguments{
\item{address}{A character string with no white space which is used as the sample address}
}
\description{
This functions invokes a system command to \code{Zint} and outputs a \strong{Code 128} barcode.
}
\details{
Barcodes are saved to the directory which has been created using \code{create_directory}.
}
