% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_barcodes.R
\name{generate_barcodes}
\alias{generate_barcodes}
\title{Generate Barcodes}
\usage{
generate_barcodes(filePath, ID, DOB, GENDER, ADDRESS, TYPE = c("FORM",
  "LOG", "FBC"))
}
\arguments{
\item{filePath}{A valid directory path where barcodes are saved}

\item{ID}{A character string of participant ID}

\item{DOB}{A character string of Date-of Birth in the following format; \strong{DD/MM/YYYY}}

\item{GENDER}{A character of either M or F (for Male and Female respectively)}

\item{ADDRESS}{A character string with no white space which is used as the sample address}

\item{TYPE}{A character vector of any of the following options
\itemize{
\item FORM
\item LOG
\item FBC
\item CHEM21
}}
}
\description{
This is the main worker function of the whole package. \code{generate_barcodes} wraps up the barcode creation and arrangement and
saves them in the specified directory.
}
